Refactor Option and Try so they share implementations and provide an API to access the values they wrap.
