+++ README.md
@@ -2,20 +2,20 @@
 
 # shuttlecock
 
-An algebraic defined types library inspired by the course [Category theory for programmers by Bartosz Milewski](https://www.youtube.com/playlist?list=PLbgaMIhjbmEnaH_LTkxLI7FMa2HsnawM_).
+A monad library based on the standard Scala libraries and heavily influenced
+by Haskell.
 
 ## Usage
 
-A simple usage example:
-
-    import 'package:shuttlecock/shuttlecock.dart';
-
-    main() {
-      var awesome = new Awesome();
-    }
+Still in development, look for Scala and Haskell documentation, it should be
+straightforward to adapt to this.
 
 ## Features and bugs
 
 Please file feature requests and bugs at the [issue tracker][tracker].
 
-[tracker]: http://example.com/issues/replaceme
+[tracker]: https://github.com/alexeieleusis/shuttlecock/issues
+
+---
+
+Originally inspired  by the course [Category theory for programmers by Bartosz Milewski](https://www.youtube.com/playlist?list=PLbgaMIhjbmEnaH_LTkxLI7FMa2HsnawM_).
