+++ lib/shuttlecock.dart
@@ -19,3 +19,4 @@ export 'src/instances/state.dart';
 export 'src/instances/stream_monad.dart';
 export 'src/instances/try.dart';
 export 'src/instances/writer.dart';
+export 'src/util/value_wrapper.dart';
